refactor BlockchainCppModule to be as independent to load a config from diff location
