+++ apps/libs/actuator/src/BlockchainCppIndicator.ts
@@ -1,8 +1,6 @@
-import { Global, Injectable, Module } from '@nestjs/common'
-import { JsonRpcClient } from '@defichain/jellyfish-api-jsonrpc'
-import { ConfigService } from '@nestjs/config'
+import { Injectable } from '@nestjs/common'
 import { ApiClient, blockchain as bc } from '@defichain/jellyfish-api-core'
-import { ActuatorProbes, ProbeIndicator } from './ActuatorModule'
+import { ProbeIndicator } from './ActuatorModule'
 import { HealthIndicatorResult } from '@nestjs/terminus'
 
 /**
@@ -58,41 +56,5 @@ export class BlockchainCppProbeIndicator extends ProbeIndicator {
     }
 
     return this.withAlive('blockchain', details)
-  }
-}
-
-/**
- * Ain Module configures and export JsonRpcClient connected to a DeFiD.
- * This does not have any side effect it merely configures and export a JsonRpcClient.
- */
-@Global()
-@Module({
-  providers: [
-    BlockchainCppProbeIndicator,
-    {
-      provide: ApiClient,
-      useFactory: (configService: ConfigService): ApiClient => {
-        const url = configService.get<string>('BLOCKCHAIN_CPP_URL')
-        if (url === undefined) {
-          throw new Error('BlockchainCppModule config:BLOCKCHAIN_CPP_URL not provided')
-        }
-        return new JsonRpcClient(configService.get<string>('BLOCKCHAIN_CPP_URL') as string)
-      },
-      inject: [ConfigService]
-    }
-  ],
-  exports: [
-    ApiClient
-  ]
-})
-export class BlockchainCppModule {
-  constructor (
-    private readonly probes: ActuatorProbes,
-    private readonly ainProbeIndicator: BlockchainCppProbeIndicator
-  ) {
-  }
-
-  async onApplicationBootstrap (): Promise<void> {
-    this.probes.add(this.ainProbeIndicator)
   }
 }
