+++ apps/libs/actuator/src/index.ts
@@ -1,3 +1,3 @@
 export * from './ActuatorModule'
 export * from './ActuatorController'
-export * from './BlockchainCppModule'
+export * from './BlockchainCppIndicator'
