+++ apps/ocean-api/src/modules/RootModule.ts
@@ -2,7 +2,8 @@ import * as Joi from 'joi'
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { ControllerModule } from './ControllerModule'
-import { ActuatorModule, BlockchainCppModule } from '@defichain-apps/libs/actuator'
+import { BlockchainCppModule } from './BlockchainCppModule'
+import { ActuatorModule } from '@defichain-apps/libs/actuator'
 import { PlaygroundModule } from './PlaygroundModule'
 
 @Module({
